move things to group structure
