+++ web/src/app/(home)/[group]/chat/[chatid]/page.tsx
@@ -1,0 +1,100 @@
+'use server'; // since this is an async component
+import React from 'react';
+
+import { Metadata, ResolvingMetadata } from 'next';
+
+import { auth } from '@/auth';
+import MainContent from '@/components/chat/main';
+import { AI } from '@/lib/aiProvider';
+import { isWonkSuccess } from '@/lib/error/error';
+import WonkyPageError from '@/lib/error/wonkyPageError';
+import { cleanMetadataTitle } from '@/lib/util';
+import { ChatHistory, blankAIState } from '@/models/chat';
+import { getFocusWithSubFocus, focuses } from '@/models/focus';
+import { WonkSession } from '@/models/session';
+import { getChat } from '@/services/historyService';
+
+type HomePageProps = {
+  params: {
+    group: string;
+    chatid: string;
+  };
+  searchParams: {
+    focus?: string;
+    subFocus?: string;
+  };
+};
+
+const getCachedChat = React.cache(async (chatid: string) => {
+  const chat = await getChat(chatid);
+
+  return chat;
+});
+
+export async function generateMetadata(
+  { params, searchParams }: HomePageProps,
+  parent: ResolvingMetadata
+): Promise<Metadata> {
+  const { group, chatid } = params;
+
+  if (chatid === 'new') {
+    return {
+      title: 'New Chat',
+    };
+  }
+
+  const result = await getCachedChat(chatid);
+
+  return {
+    title:
+      isWonkSuccess(result) && result.data.title
+        ? cleanMetadataTitle(result.data.title)
+        : 'Chat',
+  };
+}
+
+const ChatPage = async ({
+  params: { group, chatid }, // Destructure group
+  searchParams: { focus, subFocus },
+}: HomePageProps) => {
+  let chat: ChatHistory;
+
+  if (chatid !== 'new') {
+    // any unexpected or server errors will be caught by the error.tsx boundary instead of crashing the page
+    const result = await getCachedChat(chatid);
+    if (!isWonkSuccess(result)) {
+      return <WonkyPageError status={result.status} />;
+    }
+    chat = result.data;
+  } else {
+    const session = (await auth()) as WonkSession;
+    chat = newChatSession(session, focus, subFocus);
+  }
+
+  return (
+    <AI initialAIState={chat}>
+      <MainContent />
+    </AI>
+  );
+};
+
+export default ChatPage;
+
+const newChatSession = (
+  session: WonkSession,
+  focusParam?: string,
+  subFocusParam?: string
+) => {
+  const focus = getFocusWithSubFocus(focusParam, subFocusParam);
+
+  const chat: ChatHistory = {
+    ...blankAIState,
+    // id is '' in state until submitUserMessage() is called
+    meta: {
+      focus: focus ?? focuses[0],
+    },
+    userId: session.userId,
+  };
+
+  return chat;
+};
