+++ web/src/app/(home)/[group]/page.tsx
@@ -19,6 +19,11 @@ const RedirectComponent: React.FC<RedirectComponentProps> = ({
     }
   }
 
+  console.log(
+    'Redirecting to /ucdavis/chat/new with params:',
+    params.toString()
+  );
+
   const destination = `/ucdavis/chat/new?${params.toString()}`;
   redirect(destination);
